Extract code into base class.
This will allow multiple verifiers with different logic.
