+++ spec/features/drug_alert_email_verification_spec.rb
@@ -27,14 +27,7 @@ RSpec.describe "Drug email alert verifier" do
   end
 
   def given_credentials_for_the_google_api_have_been_set
-    ENV['GOOGLE_OAUTH_CREDENTIALS'] = '{"client_id":"my-google-client-id","access_token":"my-access-token","refresh_token":"my-refresh-token","scope":["https://www.googleapis.com/auth/gmail.readonly"],"expiration_time_millis":1454336608000}'
-    ENV['GOOGLE_CLIENT_ID'] = 'my-google-client-id'
-    ENV['GOOGLE_CLIENT_SECRET'] = 'my-google-client-secret'
-    ENV['EMAILS_THAT_SHOULD_RECEIVE_DRUG_ALERTS'] = 'a@example.org,b@example.org'
-
-    # the response there doesn't matter, as long as it's JSON.
-    stub_request(:post, "https://www.googleapis.com/oauth2/v3/token").
-      to_return(body: "{}", headers: { 'Content-Type' => 'application/json'})
+    set_credentials
   end
 
   def and_there_are_drug_advice_alerts
